done with ajax workshop

+++ public/scripts/Hotel.js
@@ -16,21 +16,29 @@ $(document).ready(function () {
 				.eraseMarker()
 				.eraseItineraryItem();
 			currentDay.hotel = null;
+			$.ajax({
+				type: 'DELETE',
+				url: '/days/' + currentDay._id + '/hotel'
+			});
 		};
 	});
 	// construct a new hotel for the current day
-	Hotel = function (data) {
+	Hotel = function (data, day_num) {
 		var self = this;
 		eachKeyValue(data, function (key, val) {
 			self[key] = val;
 		});
-		if (currentDay.hotel) {
+		if (currentDay && currentDay.hotel) {
 			currentDay.hotel.delete();
 		}
 		this.buildMarker()
-			.drawMarker()
-			.buildItineraryItem()
-			.drawItineraryItem();
-		currentDay.hotel = this;
-	}	
+			.buildItineraryItem();
+
+		if (day_num == undefined) {
+			this.drawMarker()
+				.drawItineraryItem();
+			currentDay.hotel = this;
+			$.post('/days/' + currentDay._id + '/hotel',{hotel: self._id});
+		}
+	}
 });
